+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## 1.0.0 (2024-03-22)
 
-
 ### Features
 
-* Web3 RPC Handler (npm package) ([#1](https://github.com/ubiquity/rpc-handler/issues/1)) ([#4](https://github.com/ubiquity/rpc-handler/issues/4)) ([0d964b0](https://github.com/ubiquity/rpc-handler/commit/0d964b09871b09fa5ce441d3339254d171ecae62))
-* workflow ([300603c](https://github.com/ubiquity/rpc-handler/commit/300603cbdc788b44c80a366021142f99842590ba))
+- Web3 RPC Handler (npm package) ([#1](https://github.com/ubiquity/rpc-handler/issues/1)) ([#4](https://github.com/ubiquity/rpc-handler/issues/4)) ([0d964b0](https://github.com/ubiquity/rpc-handler/commit/0d964b09871b09fa5ce441d3339254d171ecae62))
+- workflow ([300603c](https://github.com/ubiquity/rpc-handler/commit/300603cbdc788b44c80a366021142f99842590ba))